refactor(mind): extract abstract node transforms function to transforms/abstract-node
