+++ .changeset/blue-buttons-sell.md
@@ -1,0 +1,5 @@
+---
+'@plait/mind': patch
+---
+
+extract abstract node transforms function to transforms/abstract-node
